Refactor Parser class into a package
